pyproject.toml
[tool.poetry]
name = "homeowner-maintanence-bot"
version = "0.0.0"
description = "A bot that keeps track of your homeowner chores."
authors = [
    "Richard Ni <richardjni96@gmail.com>",
]
readme = "README.md"

packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8.2"
flask = "^2.3.3"
gunicorn = "^21.2.0"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.20"
mysql-connector-python = "^8.1.0"
icecream = "^2.1.3"

[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[tool.poetry.group.test.dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120

[tool.isort]
profile = "black"
line_length = 120

[tool.flake8]
max-line-length = 120
inline-quotes = "double"
